refactor: search bar to app layout
